Refactorize ZXBPP
Move it into its own directory along with its dependencies
