+++ parsetab/tabs.dbm.dir
@@ -1,4 +1,4 @@
-'zxbppparse', (0, 68810)
-'asmparse', (69120, 250690)
-'zxnext_asmparse', (320000, 281648)
-'zxbparser', (602112, 707013)
+'zxbppparse', (0, 69188)
+'asmparse', (69632, 250690)
+'zxnext_asmparse', (320512, 281648)
+'zxbparser', (602624, 707013)
