+++ zxbpp/__init__.py
@@ -1,0 +1,1 @@
+from .zxbpp import entry_point  # noqa
