+++ zxbpp/prepro/definestable.py
